+++ packages/dataviews/src/hooks/use-form-validity.ts
@@ -13,9 +13,8 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import normalizeFields from '../utils/normalize-fields';
-import normalizeFormFields from '../dataform-layouts/normalize-form-fields';
+import normalizeForm from '../dataform-layouts/normalize-form';
 import type {
-	CombinedFormField,
 	Field,
 	FieldValidity,
 	Form,
@@ -111,16 +110,16 @@ function getFieldsToValidate< Item >(
 	fields: NormalizedField< Item >[];
 	fieldToParent: Map< string, string >;
 } {
-	const formFields = normalizeFormFields( form );
-	if ( formFields.length === 0 ) {
+	const normalizedForm = normalizeForm( form );
+	if ( normalizedForm.fields.length === 0 ) {
 		return { fields: [], fieldToParent: new Map() };
 	}
 
 	const fieldToParent = new Map< string, string >();
 	const fieldIdsToValidate: string[] = [];
-	formFields.forEach( ( formField ) => {
-		if ( !! ( formField as CombinedFormField ).children ) {
-			( formField as CombinedFormField ).children.forEach( ( child ) => {
+	normalizedForm.fields.forEach( ( formField ) => {
+		if ( !! formField.children ) {
+			formField.children.forEach( ( child ) => {
 				const childId = typeof child === 'string' ? child : child.id;
 				fieldIdsToValidate.push( childId );
 				fieldToParent.set( childId, formField.id );